backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.25.2
Pillow>=12.0.0  # 图片处理
python-multipart==0.0.6  # 文件上传